Added single product page

+++ pages/api/products/[id].js
@@ -7,33 +7,36 @@ export default async function handler(req, res) {
     query: { id },
   } = req;
 
-  connectDB();
+  await connectDB();
 
-  switch (method) {
-    case 'GET':
-      try {
-        const product = await Product.findById(id);
-        res.status(200).json(product);
-      } catch (error) {
-        res.status(500).json(error);
-      }
-    case 'DELETE':
-      try {
-        await Product.findByIdAndDelete(id);
-        res.status(200).json({
-          message: 'Product has been deleted',
-        });
-      } catch (error) {
-        res.status(500).json(error);
-      }
-    case 'PUT':
-      try {
-        const product = await Product.findByIdAndUpdate(id, req.body, {
-          new: true,
-        });
-        res.status(200).json(product);
-      } catch (error) {
-        res.status(500).json(error);
-      }
+  if (method === 'GET') {
+    try {
+      const product = await Product.findById(id);
+      res.status(200).json(product);
+    } catch (error) {
+      res.status(500).json(error);
+    }
+  }
+
+  if (method === 'DELETE') {
+    try {
+      await Product.findByIdAndDelete(id);
+      res.status(200).json({
+        message: 'Product has been deleted',
+      });
+    } catch (error) {
+      res.status(500).json(error);
+    }
+  }
+
+  if (method === 'PUT') {
+    try {
+      const product = await Product.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+      res.status(200).json(product);
+    } catch (error) {
+      res.status(500).json(error);
+    }
   }
 }
